requirements.txt
speechrecognition
pyttsx3
google-generativeai
python-dotenv
webbrowser
